Split HackyAI into a more manageable set of files.
