+++ OpenRA.Mods.RA/AI/States/GroundStates.cs
@@ -1,0 +1,172 @@
+#region Copyright & License Information
+/*
+ * Copyright 2007-2013 The OpenRA Developers (see AUTHORS)
+ * This file is part of OpenRA, which is free software. It is made
+ * available to you under the terms of the GNU General Public License
+ * as published by the Free Software Foundation. For more information,
+ * see COPYING.
+ */
+#endregion
+
+using System.Collections.Generic;
+using System.Linq;
+using OpenRA.Traits;
+
+namespace OpenRA.Mods.RA.AI
+{
+	abstract class GroundStateBase : StateBase
+	{
+		protected virtual bool MayBeFlee(Squad owner)
+		{
+			return base.MayBeFlee(owner, (enemyAroundUnit) =>
+			{
+				owner.attackOrFleeFuzzy.CalculateFuzzy(owner.units, enemyAroundUnit);
+				if (!owner.attackOrFleeFuzzy.CanAttack)
+					return true;
+
+				return false;
+			});
+		}
+	}
+
+	class GroundUnitsIdleState : GroundStateBase, IState
+	{
+		public void Enter(Squad owner) { }
+
+		public void Execute(Squad owner)
+		{
+			if (owner.IsEmpty) return;
+			if (!owner.TargetIsValid)
+			{
+				var t = owner.bot.FindClosestEnemy(owner.units.FirstOrDefault().CenterPosition);
+				if (t == null) return;
+				owner.Target = t;
+			}
+
+			var enemyUnits = owner.world.FindActorsInCircle(owner.Target.CenterPosition, WRange.FromCells(10))
+				.Where(unit => owner.bot.p.Stances[unit.Owner] == Stance.Enemy).ToList();
+			if (enemyUnits.Any())
+
+			{
+				owner.attackOrFleeFuzzy.CalculateFuzzy(owner.units, enemyUnits);
+				if (owner.attackOrFleeFuzzy.CanAttack)
+				{
+					foreach(var u in owner.units)
+						owner.world.IssueOrder(new Order("AttackMove", u, false) { TargetLocation = owner.Target.CenterPosition.ToCPos() });
+					// We have gathered sufficient units. Attack the nearest enemy unit.
+					owner.fsm.ChangeState(new GroundUnitsAttackMoveState(), true);
+					return;
+				}
+				else
+					owner.fsm.ChangeState(new GroundUnitsFleeState(), true);
+			}
+		}
+
+		public void Exit(Squad owner) { }
+	}
+
+	class GroundUnitsAttackMoveState : GroundStateBase, IState
+	{
+		public void Enter(Squad owner) { }
+
+		public void Execute(Squad owner)
+		{
+			if (owner.IsEmpty) return;
+
+			if (!owner.TargetIsValid)
+			{
+				var closestEnemy = owner.bot.FindClosestEnemy(owner.units.Random(owner.random).CenterPosition);
+				if (closestEnemy != null)
+					owner.Target = closestEnemy;
+				else
+				{
+					owner.fsm.ChangeState(new GroundUnitsFleeState(), true);
+					return;
+				}
+			}
+
+			Actor leader = owner.units.ClosestTo(owner.Target.CenterPosition);
+			if (leader == null)
+				return;
+			var ownUnits = owner.world.FindActorsInCircle(leader.CenterPosition, WRange.FromCells(owner.units.Count) / 3)
+				.Where(a => a.Owner == owner.units.FirstOrDefault().Owner && owner.units.Contains(a)).ToList();
+			if (ownUnits.Count < owner.units.Count)
+			{
+				owner.world.IssueOrder(new Order("Stop", leader, false));
+				foreach (var unit in owner.units.Where(a => !ownUnits.Contains(a)))
+					owner.world.IssueOrder(new Order("AttackMove", unit, false) { TargetLocation = leader.CenterPosition.ToCPos() });
+			}
+			else
+			{
+				var enemys = owner.world.FindActorsInCircle(leader.CenterPosition, WRange.FromCells(12))
+					.Where(a1 => !a1.Destroyed && !a1.IsDead()).ToList();
+				var enemynearby = enemys.Where(a1 => a1.HasTrait<ITargetable>() && leader.Owner.Stances[a1.Owner] == Stance.Enemy).ToList();
+				if (enemynearby.Any())
+				{
+					owner.Target = enemynearby.ClosestTo(leader.CenterPosition);
+					owner.fsm.ChangeState(new GroundUnitsAttackState(), true);
+					return;
+				}
+				else
+					foreach (var a in owner.units)
+						owner.world.IssueOrder(new Order("AttackMove", a, false) { TargetLocation = owner.Target.Location });
+			}
+
+			if (MayBeFlee(owner))
+			{
+				owner.fsm.ChangeState(new GroundUnitsFleeState(), true);
+				return;
+			}
+		}
+
+		public void Exit(Squad owner) { }
+	}
+
+	class GroundUnitsAttackState : GroundStateBase, IState
+	{
+		public void Enter(Squad owner) { }
+
+		public void Execute(Squad owner)
+		{
+			if (owner.IsEmpty) return;
+
+			if (!owner.TargetIsValid)
+			{
+				var closestEnemy = owner.bot.FindClosestEnemy(owner.units.Random(owner.random).CenterPosition);
+				if (closestEnemy != null)
+					owner.Target = closestEnemy;
+				else
+				{
+					owner.fsm.ChangeState(new GroundUnitsFleeState(), true);
+					return;
+				}
+			}
+			foreach (var a in owner.units)
+				if (!BusyAttack(a))
+					owner.world.IssueOrder(new Order("Attack", a, false) { TargetActor = owner.bot.FindClosestEnemy(a.CenterPosition) });
+
+			if (MayBeFlee(owner))
+			{
+				owner.fsm.ChangeState(new GroundUnitsFleeState(), true);
+				return;
+			}
+		}
+
+		public void Exit(Squad owner) { }
+	}
+
+	class GroundUnitsFleeState : GroundStateBase, IState
+	{
+		public void Enter(Squad owner) { }
+
+		public void Execute(Squad owner)
+		{
+			if (owner.IsEmpty) return;
+
+			GoToRandomOwnBuilding(owner);
+			owner.fsm.ChangeState(new GroundUnitsIdleState(), true);
+		}
+
+		public void Exit(Squad owner) { owner.units.Clear(); }
+	}
+}
